pyproject.toml
[project]
name = "sig"
version = "0.1.0"
description = "Paper-aligned SRDiff evaluation code for spectral-aligned inference guidance"
requires-python = ">=3.10,<3.11"
dependencies = [
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "lpips>=0.1.4",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pytorch-msssim>=1.0.0",
    "pyyaml>=6.0.3",
    "scikit-image>=0.25.2",
    "torch>=2.5.1",
    "torch-fidelity>=0.3.0",
    "torchmetrics>=1.8.2",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
]

[tool.uv]
package = false

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
